requirements.txt
Django==3.0.3
awesome-slugify==1.6.5
Pillow==7.2.0
python-dotenv==0.15.0
psycopg2-binary==2.8.5
gunicorn==20.0.4
